Expanded the container definition model
The existing configuration model was incomplete and the code that consumed it didn't allow for adding additional values. Generated an updated model from the ECS JSON Schema to ensure that it is accurate and up to date and then modified the code that consumes it to generate the appropriate JSON object by exporting the model object directly, allowing for the use of field aliases to ensure the proper naming.